Add pytorch implementation for example othello game
