+++ example/othello/keras.py
@@ -4,6 +4,17 @@ import numpy as np
 from alpha_zero_general import DotDict
 from alpha_zero_general import NeuralNet
 
+from tensorflow.keras.layers import Activation
+from tensorflow.keras.layers import BatchNormalization
+from tensorflow.keras.layers import Conv2D
+from tensorflow.keras.layers import Dense
+from tensorflow.keras.layers import Dropout
+from tensorflow.keras.layers import Flatten
+from tensorflow.keras.layers import Input
+from tensorflow.keras.layers import Reshape
+from tensorflow.keras.models import Model
+from tensorflow.keras.optimizers import Adam
+
 args = DotDict(
     {
         "lr": 0.001,
@@ -76,3 +87,75 @@ class KerasNetWrapper(NeuralNet):
         # https://github.com/pytorch/examples/blob/master/imagenet/main.py#L98
         filepath = os.path.join(folder, filename)
         self.model.load_weights(filepath)
+
+
+class OthelloNNet(KerasNetWrapper):
+    @staticmethod
+    def get_model(board_size, action_size, args):
+        # game params
+        board_x, board_y = board_size
+        action_size = action_size
+
+        # Neural Net
+
+        # s: batch_size x board_x x board_y
+        input_boards = Input(shape=(board_x, board_y))
+        # batch_size  x board_x x board_y x 1
+        x_image = Reshape((board_x, board_y, 1))(input_boards)
+        # batch_size  x board_x x board_y x num_channels
+        h_conv1 = Activation("relu")(
+            BatchNormalization(axis=3)(
+                Conv2D(args.num_channels, 3, padding="same", use_bias=False)(
+                    x_image
+                )
+            )
+        )
+        # batch_size  x board_x x board_y x num_channels
+        h_conv2 = Activation("relu")(
+            BatchNormalization(axis=3)(
+                Conv2D(args.num_channels, 3, padding="same", use_bias=False)(
+                    h_conv1
+                )
+            )
+        )
+        # batch_size  x (board_x-2) x (board_y-2) x num_channels
+        h_conv3 = Activation("relu")(
+            BatchNormalization(axis=3)(
+                Conv2D(args.num_channels, 3, padding="valid", use_bias=False)(
+                    h_conv2
+                )
+            )
+        )
+        # batch_size  x (board_x-4) x (board_y-4) x num_channels
+        h_conv4 = Activation("relu")(
+            BatchNormalization(axis=3)(
+                Conv2D(args.num_channels, 3, padding="valid", use_bias=False)(
+                    h_conv3
+                )
+            )
+        )
+        h_conv4_flat = Flatten()(h_conv4)
+        # batch_size x 1024
+        s_fc1 = Dropout(args.dropout)(
+            Activation("relu")(
+                BatchNormalization(axis=1)(
+                    Dense(1024, use_bias=False)(h_conv4_flat)
+                )
+            )
+        )
+        # batch_size x 1024
+        s_fc2 = Dropout(args.dropout)(
+            Activation("relu")(
+                BatchNormalization(axis=1)(Dense(512, use_bias=False)(s_fc1))
+            )
+        )
+        # batch_size x action_size
+        pi = Dense(action_size, activation="softmax", name="pi")(s_fc2)
+        # batch_size x 1
+        v = Dense(1, activation="tanh", name="v")(s_fc2)
+        model = Model(inputs=input_boards, outputs=[pi, v])
+        model.compile(
+            loss=["categorical_crossentropy", "mean_squared_error"],
+            optimizer=Adam(args.lr),
+        )
+        return model
